lib_project/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
drf-yasg
gunicorn
django-cors-headers
